requirements.txt
Flask==3.0.3
selenium==4.23.1
gspread==6.1.2
oauth2client==4.1.3
python-dateutil==2.9.0.post0